Pulled chromium 77.0.3865.92

+++ build/config/ios/codesign.py
@@ -243,12 +243,15 @@ def InstallSystemFramework(framework_path, bundle_path, args):
   installed_framework_path = os.path.join(
       bundle_path, 'Frameworks', os.path.basename(framework_path))
 
-  if os.path.exists(installed_framework_path):
-    shutil.rmtree(installed_framework_path)
+  if os.path.isfile(framework_path):
+    shutil.copy(framework_path, installed_framework_path)
+  elif os.path.isdir(framework_path):
+    if os.path.exists(installed_framework_path):
+      shutil.rmtree(installed_framework_path)
+    shutil.copytree(framework_path, installed_framework_path)
 
-  shutil.copytree(framework_path, installed_framework_path)
   CodeSignBundle(installed_framework_path, args.identity,
-      ['--deep', '--preserve-metadata=identifier,entitlements'])
+      ['--deep', '--preserve-metadata=identifier,entitlements,flags'])
 
 
 def GenerateEntitlements(path, provisioning_profile, bundle_identifier):
@@ -508,8 +511,6 @@ class GenerateEntitlementsAction(Action):
 
 def Main():
   parser = argparse.ArgumentParser('codesign iOS bundles')
-  parser.add_argument('--developer_dir', required=False,
-                      help='Path to Xcode.')
   subparsers = parser.add_subparsers()
 
   actions = [
@@ -522,8 +523,6 @@ def Main():
     action.Register(subparsers)
 
   args = parser.parse_args()
-  if args.developer_dir:
-    os.environ['DEVELOPER_DIR'] = args.developer_dir
   args.func(args)
