+++ chrome/browser/notifications/notification_channels_provider_android.cc
@@ -6,15 +6,19 @@
 
 #include <algorithm>
 
+#include "base/android/build_info.h"
 #include "base/android/jni_android.h"
 #include "base/android/jni_string.h"
+#include "base/bind.h"
 #include "base/feature_list.h"
 #include "base/logging.h"
 #include "base/macros.h"
 #include "base/strings/string_number_conversions.h"
 #include "base/strings/string_util.h"
+#include "base/task/post_task.h"
 #include "base/time/default_clock.h"
 #include "base/values.h"
+#include "chrome/android/chrome_jni_headers/NotificationSettingsBridge_jni.h"
 #include "chrome/browser/content_settings/host_content_settings_map_factory.h"
 #include "chrome/common/chrome_features.h"
 #include "chrome/common/pref_names.h"
@@ -27,13 +31,16 @@
 #include "components/content_settings/core/common/content_settings_pattern.h"
 #include "components/content_settings/core/common/content_settings_utils.h"
 #include "components/prefs/scoped_user_pref_update.h"
+#include "components/search_engines/template_url.h"
+#include "components/search_engines/template_url_service.h"
+#include "content/public/browser/browser_task_traits.h"
 #include "content/public/browser/browser_thread.h"
-#include "jni/NotificationSettingsBridge_jni.h"
 #include "url/gurl.h"
 #include "url/origin.h"
 #include "url/url_constants.h"
 
 using base::android::AttachCurrentThread;
+using base::android::BuildInfo;
 using base::android::ConvertUTF8ToJavaString;
 using base::android::ScopedJavaLocalRef;
 
@@ -46,8 +53,8 @@ class NotificationChannelsBridgeImpl
   ~NotificationChannelsBridgeImpl() override = default;
 
   bool ShouldUseChannelSettings() override {
-    return Java_NotificationSettingsBridge_shouldUseChannelSettings(
-        AttachCurrentThread());
+    return BuildInfo::GetInstance()->sdk_int() >=
+           base::android::SDK_VERSION_OREO;
   }
 
   NotificationChannel CreateChannel(const std::string& origin,
@@ -85,11 +92,8 @@ class NotificationChannelsBridgeImpl
     JNIEnv* env = AttachCurrentThread();
     ScopedJavaLocalRef<jobjectArray> raw_channels =
         Java_NotificationSettingsBridge_getSiteChannels(env);
-    jsize num_channels = env->GetArrayLength(raw_channels.obj());
     std::vector<NotificationChannel> channels;
-    for (jsize i = 0; i < num_channels; ++i) {
-      ScopedJavaLocalRef<jobject> jchannel(
-          env, env->GetObjectArrayElement(raw_channels.obj(), i));
+    for (auto jchannel : raw_channels.ReadElements<jobject>()) {
       channels.push_back(NotificationChannel(
           ConvertJavaStringToUTF8(Java_SiteChannel_getId(env, jchannel)),
           ConvertJavaStringToUTF8(Java_SiteChannel_getOrigin(env, jchannel)),
@@ -130,8 +134,7 @@ class ChannelsRuleIterator : public content_settings::RuleIterator {
         ContentSettingsPattern::FromURLNoWildcard(
             GURL(channels_[index_].origin)),
         ContentSettingsPattern::Wildcard(),
-        new base::Value(
-            ChannelStatusToContentSetting(channels_[index_].status)));
+        base::Value(ChannelStatusToContentSetting(channels_[index_].status)));
     index_++;
     return rule;
   }
@@ -142,11 +145,33 @@ class ChannelsRuleIterator : public content_settings::RuleIterator {
   DISALLOW_COPY_AND_ASSIGN(ChannelsRuleIterator);
 };
 
+// This copies the logic of
+// SearchPermissionsService::IsPermissionControlledByDSE, which cannot be
+// called from this class as it would introduce a circular dependency between
+// the HostContentSettingsMap and the SearchPermissionsService factories.
+bool OriginMatchesDefaultSearchEngine(TemplateURLService* template_url_service,
+                                      const std::string& origin) {
+  if (!template_url_service)
+    return false;
+
+  const TemplateURL* default_search_engine =
+      template_url_service->GetDefaultSearchProvider();
+
+  if (!default_search_engine)
+    return false;
+
+  GURL default_search_engine_url = default_search_engine->GenerateSearchURL(
+      template_url_service->search_terms_data());
+
+  return url::IsSameOriginWith(GURL(origin), default_search_engine_url);
+}
 }  // anonymous namespace
 
 // static
 void NotificationChannelsProviderAndroid::RegisterProfilePrefs(
     user_prefs::PrefRegistrySyncable* registry) {
+  registry->RegisterBooleanPref(prefs::kClearedBlockedSiteNotificationChannels,
+                                false /* default_value */);
   registry->RegisterBooleanPref(prefs::kMigratedToSiteNotificationChannels,
                                 false);
 }
@@ -196,9 +221,8 @@ void NotificationChannelsProviderAndroid::MigrateToChannelsIfNecessary(
 
     while (it && it->HasNext()) {
       content_settings::Rule rule = it->Next();
-      rules.push_back(rule);
-
       CreateChannelForRule(rule);
+      rules.push_back(std::move(rule));
     }
   }
 
@@ -213,29 +237,30 @@ void NotificationChannelsProviderAndroid::MigrateToChannelsIfNecessary(
   prefs->SetBoolean(prefs::kMigratedToSiteNotificationChannels, true);
 }
 
-void NotificationChannelsProviderAndroid::UnmigrateChannelsIfNecessary(
+void NotificationChannelsProviderAndroid::ClearBlockedChannelsIfNecessary(
     PrefService* prefs,
-    content_settings::ProviderInterface* pref_provider) {
+    TemplateURLService* template_url_service) {
   if (!platform_supports_channels_ ||
-      !prefs->GetBoolean(prefs::kMigratedToSiteNotificationChannels)) {
+      prefs->GetBoolean(prefs::kClearedBlockedSiteNotificationChannels)) {
     return;
   }
-  std::unique_ptr<content_settings::RuleIterator> it(
-      GetRuleIterator(CONTENT_SETTINGS_TYPE_NOTIFICATIONS, std::string(),
-                      false /* incognito */));
-  while (it && it->HasNext()) {
-    const content_settings::Rule& rule = it->Next();
-    pref_provider->SetWebsiteSetting(rule.primary_pattern,
-                                     rule.secondary_pattern,
-                                     CONTENT_SETTINGS_TYPE_NOTIFICATIONS,
-                                     content_settings::ResourceIdentifier(),
-                                     new base::Value(rule.value->Clone()));
-  }
-  for (auto& channel : bridge_->GetChannels())
-    bridge_->DeleteChannel(channel.id);
-  cached_channels_.clear();
 
-  prefs->SetBoolean(prefs::kMigratedToSiteNotificationChannels, false);
+  for (const NotificationChannel& channel : bridge_->GetChannels()) {
+    if (channel.status != NotificationChannelStatus::BLOCKED)
+      continue;
+    if (OriginMatchesDefaultSearchEngine(template_url_service,
+                                         channel.origin)) {
+      // Do not clear the DSE permission, as it should always be ALLOW or BLOCK.
+      continue;
+    }
+    bridge_->DeleteChannel(channel.id);
+  }
+
+  // Reset the cache.
+  cached_channels_.clear();
+  initialized_cached_channels_ = false;
+
+  prefs->SetBoolean(prefs::kClearedBlockedSiteNotificationChannels, true);
 }
 
 std::unique_ptr<content_settings::RuleIterator>
@@ -264,7 +289,7 @@ NotificationChannelsProviderAndroid::UpdateCachedChannels() const {
     // underlying state of NotificationChannelsProviderAndroid, and allows us to
     // notify observers as soon as we detect changes to channels.
     auto* provider = const_cast<NotificationChannelsProviderAndroid*>(this);
-    content::BrowserThread::GetTaskRunnerForThread(content::BrowserThread::UI)
+    base::CreateSingleThreadTaskRunnerWithTraits({content::BrowserThread::UI})
         ->PostTask(FROM_HERE,
                    base::BindOnce(
                        &NotificationChannelsProviderAndroid::NotifyObservers,
@@ -282,7 +307,7 @@ bool NotificationChannelsProviderAndroid::SetWebsiteSetting(
     const ContentSettingsPattern& secondary_pattern,
     ContentSettingsType content_type,
     const content_settings::ResourceIdentifier& resource_identifier,
-    base::Value* value) {
+    std::unique_ptr<base::Value>&& value) {
   if (content_type != CONTENT_SETTINGS_TYPE_NOTIFICATIONS ||
       !platform_supports_channels_) {
     return false;
@@ -297,10 +322,10 @@ bool NotificationChannelsProviderAndroid::SetWebsiteSetting(
   InitCachedChannels();
 
   url::Origin origin = url::Origin::Create(GURL(primary_pattern.ToString()));
-  DCHECK(!origin.unique());
+  DCHECK(!origin.opaque());
   const std::string origin_string = origin.Serialize();
 
-  ContentSetting setting = content_settings::ValueToContentSetting(value);
+  ContentSetting setting = content_settings::ValueToContentSetting(value.get());
   switch (setting) {
     case CONTENT_SETTING_ALLOW:
       CreateChannelIfRequired(origin_string,
@@ -323,6 +348,7 @@ bool NotificationChannelsProviderAndroid::SetWebsiteSetting(
       NOTREACHED();
       break;
   }
+  value.reset();
   return true;
 }
 
@@ -357,7 +383,7 @@ base::Time NotificationChannelsProviderAndroid::GetWebsiteSettingLastModified(
     return base::Time();
   }
   url::Origin origin = url::Origin::Create(GURL(primary_pattern.ToString()));
-  if (origin.unique())
+  if (origin.opaque())
     return base::Time();
   const std::string origin_string = origin.Serialize();
 
@@ -373,8 +399,6 @@ base::Time NotificationChannelsProviderAndroid::GetWebsiteSettingLastModified(
 void NotificationChannelsProviderAndroid::CreateChannelIfRequired(
     const std::string& origin_string,
     NotificationChannelStatus new_channel_status) {
-  // TODO(awdf): Maybe check cached incognito status here to make sure
-  // channels are never created in incognito mode.
   auto channel_entry = cached_channels_.find(origin_string);
   if (channel_entry == cached_channels_.end()) {
     base::Time timestamp = clock_->Now();
@@ -398,10 +422,10 @@ void NotificationChannelsProviderAndroid::CreateChannelForRule(
     const content_settings::Rule& rule) {
   url::Origin origin =
       url::Origin::Create(GURL(rule.primary_pattern.ToString()));
-  DCHECK(!origin.unique());
+  DCHECK(!origin.opaque());
   const std::string origin_string = origin.Serialize();
   ContentSetting content_setting =
-      content_settings::ValueToContentSetting(rule.value.get());
+      content_settings::ValueToContentSetting(&rule.value);
   switch (content_setting) {
     case CONTENT_SETTING_ALLOW:
       CreateChannelIfRequired(origin_string,
